requirements.txt
Django==5.0.4
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-crontab==0.7.1
Pillow==10.2.0
OpenAI